adds random AI and functinality to the end game button
This commit does several things. First, it adds the random AI option, where the player plays against bot that randomly chooses what target to press. At this
point the AI cannot differenciate btw a complete smallBoard and an uncomplete
smallBoard. The other update at this point adds functionality to the stop
button. It ends that game by clearing the board and reseting all of the
variables.