fixes AI to see completed boards
This commit fixes the bug where the AI could not detect the completed boards.
This means that if the player were to choose a spot that correlates with
a completed board, the AI will now randomly choose a smallBoard aswell
as a random button. Originally, it would still put it answer in the
completed board.