adds a win page and almost done
This commit adds the win page that puts up a big icon when
a player wins. This finishes off pretty much everything, with the
expetion of the title page. All of the game UI and logic is in place.
Now all is left is writing instructions and adding comments in the
code for readability.